backend/requirements.txt
fastapi[standard]
uvicorn[standard]
motor
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pymongo
langchain-chroma
langchain
langchain-community
sentence-transformers
opencv-python
torch
torchvision
pillow
numpy
python-dotenv
google-generativeai
transnetv2_pytorch
open-clip-torch
transformers
scipy
langchain-huggingface